Move create_app to startup_utils
Mirror the structure of OWS
so cubedash-gen does not have
to import _model.py on startup.

Refs #1100
