+++ cubedash/startup_utils/__init__.py
@@ -1,0 +1,3 @@
+from .._model import create_app
+
+__all__ = ["create_app"]
